refactor: create seprate middleware for authinticate external service
