+++ constants/logs.ts
@@ -6,6 +6,8 @@ const logType = {
   EVENTS_REMOVE_PEER: "EVENTS_REMOVE_PEER",
   APPLICATION_ADDED: "USER_APPLICATION_ADDED",
   TASKS_MISSED_UPDATES_ERRORS: "TASKS_MISSED_UPDATES_ERRORS",
+  DISCORD_INVITES: "DISCORD_INVITES",
+  EXTERNAL_SERVICE: "EXTERNAL_SERVICE",
 };
 
 module.exports = { logType };
