+++ routes/invites.ts
@@ -2,8 +2,9 @@ import express from "express";
 const router = express.Router();
 import { createInviteValidator } from "../middlewares/validators/invites";
 import { createInviteLink,getInviteLink } from "../controllers/invites";
+import authinticateServiceRequest from "../middlewares/authinticateServiceRequest";
 
-router.post("/", createInviteValidator, createInviteLink);
-router.get("/:userId" , getInviteLink);
+router.post("/",authinticateServiceRequest, createInviteValidator, createInviteLink);
+router.get("/:userId" ,authinticateServiceRequest, getInviteLink);
 
 module.exports = router;
